requirements.txt
certifi==2024.8.30
chardet==3.0.4
et_xmlfile==2.0.0
fire==0.7.0
fonttools==4.55.2
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
lxml==5.3.0
numpy==2.2.0
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
pdf2docx==0.5.8
pillow==11.0.0
PyMuPDF==1.25.0
pypandoc==1.14
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2024.2
